+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log (vs-deploy)
 
+## 4.9.0 (January 2th, 2017; validate remote files)
+
+* can define a script for a [deploy host](https://github.com/mkloubert/vs-deploy/wiki#validation-) now, which validates a received file
+
+
 ## 4.8.0 (January 2th, 2017; deploy hosts)
 
 * optimized [deploy host](https://github.com/mkloubert/vs-deploy/wiki#host-) feature